se adjunto el log en cotizaciones

+++ frontend/renderer/schemas/cart-item.schema.js
@@ -14,7 +14,7 @@ export const cartItemSchema = z.object({
   code:      z.string().min(1),
   name:      z.string().min(1),
   price:     z.number().nonnegative(),
-  qty:       z.number().int().positive(),
+  qty:       z.number().positive(),
   maxStock:  z.number().int(), // para validar al incrementar
 })
